+++ src/test/java/pages/BillingAddressPage.java
@@ -6,7 +6,6 @@ import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 
 import static java.lang.String.format;
-import static utils.Utils.switchToIframe;
 
 public class BillingAddressPage extends BasePage {
 
@@ -25,13 +24,7 @@ public class BillingAddressPage extends BasePage {
     private final String CART_NUMBER = "//input[@id='cardNumber']";
     private final String ACCEPT_TERMS_CHECKBOX = "//input[@id='wp-CheckboxConfirmTermsAndConditions']";
     private final String PROCEED_TO_PAYMENT = "//a[@class='button primary expand btn-shipping-address-js']";
-    private final String BILLING_INFO = "//div[@class='billing-address-display']";
-    private final String MONTH_DATA_CARD = "//input[@class='textbox expiry expiry-month']";
-    private final String EXPIRY_YEAR_CARD = "//input[@class='textbox expiry expiry-year']";
-    private final String SECURITY_CODE_BUTTON = "//input[@class='textbox pin security-code']";
-    private final String CARD_HOLDER_NAME ="//input[@class='textbox cardholder-name']";
-    private final String MAKE_PAYMENT_BUTTON = "//input[@class='button btn-make-payment']";
-    private final String IFRAME = "//iframe[@title='Payment Pages']";
+
     public static int finalPrice;
 
     public BillingAddressPage(WebDriver driver) {
@@ -99,24 +92,6 @@ public class BillingAddressPage extends BasePage {
     public void proceedPayment() {
         waitUntilElementToBeClickable(PROCEED_TO_PAYMENT).click();
     }
-
-    public void enterCartForm(String cartNumber,String cardMonthBut, String cardYearBut, String securityCode , String nameCardHolder){
-        switchToIframe(driver, IFRAME);
-        waitUntilElementIsPresent(CART_NUMBER).sendKeys(cartNumber);
-        waitUntilElementIsPresent(MONTH_DATA_CARD).sendKeys(cardMonthBut);
-        waitUntilElementIsPresent(EXPIRY_YEAR_CARD).sendKeys(cardYearBut);
-        waitUntilElementIsPresent(SECURITY_CODE_BUTTON).sendKeys(securityCode);
-        waitUntilElementIsPresent(CARD_HOLDER_NAME).sendKeys(nameCardHolder);
-    }
-
-    public boolean isMakePaymentEnabled(){
-        return isElementEnabled(MAKE_PAYMENT_BUTTON);
-    }
-
-    public String getAllBillingInfo() {
-        return waitUntilVisibilityOfElementLocated(BILLING_INFO).getText();
-    }
-
 
 }
